[Bugfix] Ship a version-aligned quick start
