+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The project follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 
 - Quick Start installs the core package from PyPI before introducing optional framework integrations.
 - Release publishing now requires a protected default-branch tag, revalidates source and exact artifact digests, uses a pinned build/audit toolchain, emits build provenance, and publishes draft-populated immutable GitHub releases.
+- The release Quick Start now runs an entry point shipped in the installed wheel instead of an example from the moving `main` branch.
 
 ## [0.1.0] - 2026-08-12
