+++ tests/unit/test_readme_quickstart.py
@@ -8,13 +8,13 @@ ROOT = Path(__file__).parents[2]
 README = ROOT / "README.md"
 
 
-def test_quickstart_keeps_source_and_example_on_same_revision():
+def test_quickstart_uses_example_shipped_with_installed_package():
     readme = README.read_text()
     quickstart = readme.split("## Quick Start", maxsplit=1)[1].split(
         "### Add resiliency to Megatron Core", maxsplit=1
     )[0]
 
-    assert "git clone https://github.com/LMResiliency/lm-resiliency.git" in quickstart
-    assert "python -m pip install -e ." in quickstart
-    assert "python examples/quickstart.py" in quickstart
-    assert '--branch "v${LM_RESILIENCY_VERSION}"' not in quickstart
+    assert "python -m pip install lm-resiliency" in quickstart
+    assert "lm-resiliency-quickstart" in quickstart
+    assert "git clone" not in quickstart
+    assert "python examples/quickstart.py" not in quickstart
